requirements.txt
PyQt5
PyQt5-Qt5
PyQt5-sip
google-generativeai
psutil